Extract RTC handler into separate module.
